refactor(cli): modularize command implementations out of main.py
Moved command implementation logic from main.py into separate modules:
- cli/evals.py: evaluation and testing commands
- cli/sessions.py: run-session command
- cli/conversations.py: convs subcommands
- cli/deployments.py: deps subcommands
- Moved run_migration_dashboard to cli/migration_cli.py

Updated main.py to lazy-load these handlers using LazyCallable, keeping its startup time optimized and imports clean. Updated tests to use correct mock paths.

TAG=agy
CONV=c637b7aa-a63d-484b-a016-c476ad4f7aaf
