+++ src/cxas_scrapi/cli/main.py
@@ -18,13 +18,8 @@ from __future__ import annotations
 
 import argparse
 import datetime
-import json
 import logging
 import os
-import subprocess
-import sys
-import time
-import uuid
 from typing import TYPE_CHECKING
 
 from cxas_scrapi.cli.insights_cli import populate_insights_parser
@@ -32,11 +27,10 @@ from cxas_scrapi.cli.resources_cli import (
     register as register_resources_subparsers,
 )
 from cxas_scrapi.cli.trace_cli import register as register_trace_subparser
-from cxas_scrapi.migration.config import DEFAULT_MODEL
+
+DEFAULT_MODEL = "gemini-3.1-flash-live"
 
 if TYPE_CHECKING:
-    import pandas as pd
-
     from cxas_scrapi.cli.app import (
         app_branch,
         app_create,
@@ -48,14 +42,35 @@ if TYPE_CHECKING:
         apps_get,
         apps_list,
     )
+    from cxas_scrapi.cli.conversations import (
+        conversations_get,
+        conversations_list,
+    )
     from cxas_scrapi.cli.create_local import handle_local_create
+    from cxas_scrapi.cli.deployments import (
+        deployments_create,
+        deployments_list,
+        deployments_promote,
+    )
+    from cxas_scrapi.cli.evals import (
+        ci_test,
+        combined_evals_report_cmd,
+        export_eval,
+        local_test,
+        push_eval,
+        run_eval,
+        test_callbacks,
+        test_single_callback,
+        test_tools,
+    )
     from cxas_scrapi.cli.llm_lint import llm_lint
+    from cxas_scrapi.cli.migration_cli import run_migration_dashboard
+    from cxas_scrapi.cli.sessions import run_session
     from cxas_scrapi.cli.versions_cli import (
         app_versions_compare,
         app_versions_list,
     )
     from cxas_scrapi.core.github import init_github_action
-    from cxas_scrapi.utils.eval_utils import EvalUtils
 else:
     from cxas_scrapi.cli.utils import LazyCallable
 
@@ -72,6 +87,9 @@ else:
         "cxas_scrapi.cli.create_local", "handle_local_create"
     )
     llm_lint = LazyCallable("cxas_scrapi.cli.llm_lint", "llm_lint")
+    run_migration_dashboard = LazyCallable(
+        "cxas_scrapi.cli.migration_cli", "run_migration_dashboard"
+    )
     app_versions_list = LazyCallable(
         "cxas_scrapi.cli.versions_cli", "app_versions_list"
     )
@@ -82,1044 +100,44 @@ else:
         "cxas_scrapi.core.github", "init_github_action"
     )
 
+    # Evals
+    export_eval = LazyCallable("cxas_scrapi.cli.evals", "export_eval")
+    push_eval = LazyCallable("cxas_scrapi.cli.evals", "push_eval")
+    run_eval = LazyCallable("cxas_scrapi.cli.evals", "run_eval")
+    combined_evals_report_cmd = LazyCallable(
+        "cxas_scrapi.cli.evals", "combined_evals_report_cmd"
+    )
+    test_tools = LazyCallable("cxas_scrapi.cli.evals", "test_tools")
+    test_callbacks = LazyCallable("cxas_scrapi.cli.evals", "test_callbacks")
+    test_single_callback = LazyCallable(
+        "cxas_scrapi.cli.evals", "test_single_callback"
+    )
+    ci_test = LazyCallable("cxas_scrapi.cli.evals", "ci_test")
+    local_test = LazyCallable("cxas_scrapi.cli.evals", "local_test")
+
+    # Sessions
+    run_session = LazyCallable("cxas_scrapi.cli.sessions", "run_session")
+
+    # Conversations
+    conversations_list = LazyCallable(
+        "cxas_scrapi.cli.conversations", "conversations_list"
+    )
+    conversations_get = LazyCallable(
+        "cxas_scrapi.cli.conversations", "conversations_get"
+    )
+
+    # Deployments
+    deployments_list = LazyCallable(
+        "cxas_scrapi.cli.deployments", "deployments_list"
+    )
+    deployments_create = LazyCallable(
+        "cxas_scrapi.cli.deployments", "deployments_create"
+    )
+    deployments_promote = LazyCallable(
+        "cxas_scrapi.cli.deployments", "deployments_promote"
+    )
+
 logger = logging.getLogger(__name__)
-
-
-def export_eval(args: argparse.Namespace) -> None:
-    """Handles the 'export' command."""
-    from cxas_scrapi.core.evaluations import Evaluations, ExportFormat
-
-    print(f"Exporting evaluation: {args.evaluation_id}")
-    # Use app_name to init client. Eval ID might be full resource name.
-    eval_client = Evaluations(app_name=args.app_name)
-
-    try:
-        format_enum = (
-            ExportFormat(args.format.lower())
-            if args.format
-            else ExportFormat.YAML
-        )
-        exported_eval = eval_client.export_evaluation(
-            args.evaluation_id,
-            output_format=format_enum,
-            output_path=args.output,
-        )
-        if args.output:
-            print(f"Evaluation exported to {args.output}")
-        else:
-            print(exported_eval)
-
-    except Exception as e:
-        print(f"Failed to export evaluation: {e}")
-        sys.exit(1)
-
-
-def run_migration_dashboard(args: argparse.Namespace) -> None:
-    """Handles the unified 'cxas migrate dfcx' command, routing to
-    non-interactive run / optimize stages or the interactive TUI dashboard.
-    """
-    from cxas_scrapi.cli.migration_cli import (
-        run_end_to_end,
-        run_resume,
-        run_stage_1,
-        run_stage_2,
-        run_stage_3,
-    )
-
-    if getattr(args, "run", False):
-        # Validate E2E requirements
-        if not (
-            getattr(args, "source_agent_id", None)
-            or getattr(args, "source_zip", None)
-        ):
-            print(
-                "Error: You must provide either --source-agent-id or "
-                "--source-zip for non-interactive --run."
-            )
-            sys.exit(1)
-        if not getattr(args, "project_id", None):
-            print(
-                "Error: Target --project-id is required for "
-                "non-interactive --run."
-            )
-            sys.exit(1)
-        if not getattr(args, "target_name", None):
-            print(
-                "Error: Target --target-name is required for "
-                "non-interactive --run."
-            )
-            sys.exit(1)
-
-        run_end_to_end(args)
-
-    elif getattr(args, "optimize", False):
-        # Validate stage requirements
-        if not getattr(args, "stage", None):
-            print(
-                "Error: You must specify a target --stage (1, 2, 3, "
-                "or resume) when using --optimize."
-            )
-            sys.exit(1)
-
-        # Set default flags
-        args.yes = True  # optimize stage is non-interactive by default
-
-        if args.stage == "1":
-            if not getattr(args, "version_label", None):
-                args.version_label = "0.0.3"
-            run_stage_1(args)
-        elif args.stage == "2":
-            if not getattr(args, "version_label", None):
-                args.version_label = "0.0.4"
-            run_stage_2(args)
-        elif args.stage == "3":
-            if not getattr(args, "version_label", None):
-                args.version_label = "0.0.5"
-            run_stage_3(args)
-        elif args.stage == "resume":
-            args.yes = False  # resume is interactive picker
-            run_resume(args)
-
-    else:
-        # Default: Interactive TUI Dashboard Mode
-        from cxas_scrapi.cli.migration_cli import MigrationCLI  # noqa: PLC0415
-        from cxas_scrapi.migration.dfcx_exporter import ConversationalAgentsAPI
-
-        dashboard = MigrationCLI()
-        cx_api = ConversationalAgentsAPI()
-        dashboard.run(default_agent_name=args.default_agent_name, cx_api=cx_api)
-
-
-def push_eval(args: argparse.Namespace) -> None:
-    """Handles the 'push-eval' command."""
-    from cxas_scrapi.core.evaluations import Evaluations
-    from cxas_scrapi.utils.eval_utils import EvalUtils
-
-    print(f"Pushing evaluation(s) from {args.file} to App: {args.app_name}")
-
-    eval_client = Evaluations(app_name=args.app_name)
-    eval_utils = EvalUtils(app_name=args.app_name)
-
-    try:
-        evals = eval_utils.load_golden_evals_from_yaml(args.file)
-        if not evals:
-            print(f"No valid evaluations found in '{args.file}'.")
-            sys.exit(1)
-
-        print(f"Parsed {len(evals)} evaluation(s). Syncing...")
-        for eval_dict in evals:
-            res = eval_client.update_evaluation(
-                evaluation=eval_dict, app_name=args.app_name
-            )
-            print(f"Pushed: '{res.display_name}' ({res.name})")
-
-        print("\nPush complete.")
-
-    except Exception as e:
-        print(f"Failed to push evaluation(s): {e}")
-        sys.exit(1)
-
-
-def wait_for_evaluation_completion(
-    eval_utils: EvalUtils,
-    old_result_ids: list[str],
-    app_name: str,
-    expected_count: int = 1,
-    timeout_seconds: int = 600,
-) -> dict[str, pd.DataFrame]:
-    """Waits for all new evaluation results to appear."""
-    import pandas as pd
-
-    print(f"Waiting for {expected_count} evaluation(s) to complete...")
-    start_time = time.time()
-    while time.time() - start_time < timeout_seconds:
-        # Fetch current evaluation results
-        try:
-            df_dict = eval_utils.evals_to_dataframe()
-            df_current = df_dict.get("summary", pd.DataFrame())
-            if df_current.empty:
-                time.sleep(5)
-                continue
-
-            # Find new runs
-            current_result_ids = set(df_current["eval_result_id"].unique())
-            new_ids = current_result_ids - old_result_ids
-
-            if new_ids and len(new_ids) >= expected_count:
-                # Wait for ALL new runs to complete
-                all_completed = True
-                completed_results = []
-                for run_id in new_ids:
-                    df_new = df_current[df_current["eval_result_id"] == run_id]
-                    exec_state = (
-                        df_new["execution_state"].iloc[0]
-                        if not df_new.empty
-                        and "execution_state" in df_new.columns
-                        else "COMPLETED"
-                    )
-
-                    if exec_state not in ("COMPLETED", "ERROR"):
-                        all_completed = False
-                        break
-
-                    # Fetch trace
-                    raw = eval_utils.eval_client.get_evaluation_result(run_id)
-                    completed_results.append(raw)
-
-                if all_completed:
-                    print(f"All {len(new_ids)} evaluations completed.")
-                    return eval_utils.evals_to_dataframe(
-                        results=completed_results
-                    )
-
-        except Exception as e:
-            print(f"Error checking evaluation status: {e}")
-
-        time.sleep(5)
-
-    print("Timeout waiting for evaluation to complete.")
-    sys.exit(1)
-
-
-def filter_metrics_and_assess(
-    df_dict_new_run: dict[str, pd.DataFrame],
-    filter_auto_metrics: bool,
-) -> bool:
-    """Assesses the evaluation run and returns True if passed,
-    False otherwise."""
-    import pandas as pd
-
-    passed = True
-
-    df_new_run = df_dict_new_run.get("summary", pd.DataFrame())
-    df_expectations = df_dict_new_run.get("expectations", pd.DataFrame())
-
-    # Standard assessment: check standard status first
-    # This might encompass semantic and hallucination metrics
-
-    num_passed = 0
-    num_failed = 0
-    num_error = 0
-    if not df_new_run.empty:
-        for _, row in df_new_run.iterrows():
-            eval_stat = str(row.get("evaluation_status", "")).upper()
-            exec_stat = str(row.get("execution_state", "")).upper()
-
-            if exec_stat in ("ERROR", "ERRORED") or eval_stat in (
-                "ERROR",
-                "ERRORED",
-            ):
-                num_error += 1
-            elif eval_stat in ("PASS", "PASSED", "✅ PASSED"):
-                num_passed += 1
-            else:
-                num_failed += 1
-
-    overall_status = (
-        "PASS"
-        if num_failed == 0 and num_error == 0 and num_passed > 0
-        else "FAIL"
-        if (num_failed > 0 or num_error > 0)
-        else "UNKNOWN"
-    )
-
-    print(f"\n--- Evaluation Status: {overall_status} ---")
-    print(f"Passed: {num_passed}")
-    print(f"Failed: {num_failed}")
-    print(f"Errored: {num_error}")
-
-    if filter_auto_metrics:
-        print(
-            "\n[Targeted Assessment] Filtering out automated LLM metrics "
-            "(semantic similarity, hallucination)."
-        )
-        print("Focusing strictly on custom expectations and tool invocation.")
-
-        if (
-            not df_expectations.empty
-            and "record_type" in df_expectations.columns
-        ):
-            expectation_rows = df_expectations[
-                df_expectations["record_type"] == "summary_expectation"
-            ]
-        else:
-            expectation_rows = pd.DataFrame()
-
-        if not expectation_rows.empty:
-            failed_expectations = expectation_rows[
-                expectation_rows["not_met_count"] > 0
-            ]
-            if not failed_expectations.empty:
-                print(
-                    f"FAILED: {len(failed_expectations)} custom expectations "
-                    "not met."
-                )
-                for _, row in failed_expectations.iterrows():
-                    print(
-                        f"  - Expectation: {row['expectation']} "
-                        f"(Met: {row['met_count']}, "
-                        f"Not Met: {row['not_met_count']})"
-                    )
-                passed = False
-            else:
-                print(
-                    f"PASSED: All {len(expectation_rows)} custom expectations "
-                    "met."
-                )
-        else:
-            print("WARNING: No custom expectations found in this evaluation.")
-            # Fallback: check basic tool execution result limit
-
-    # Strict overall pass/fail based on the server constraints
-    elif overall_status != "PASS":
-        passed = False
-
-    return passed
-
-
-def run_eval(args: argparse.Namespace) -> None:
-    """Handles the 'run' command."""
-    import pandas as pd
-
-    from cxas_scrapi.core.evaluations import Evaluations
-    from cxas_scrapi.utils.eval_utils import EvalUtils
-
-    print(f"Triggering evaluation for App: {args.app_name}")
-    eval_client = Evaluations(app_name=args.app_name)
-    eval_utils = EvalUtils(app_name=args.app_name)
-
-    # Determine which evaluations to run
-    evaluations_to_run = []
-    if args.evaluation_id:
-        evaluations_to_run.append(args.evaluation_id)
-    else:
-        # Require prefix or tags if no specific ID is given
-        if not args.display_name_prefix and not args.tags:
-            print(
-                "Error: You must provide either --evaluation-id, "
-                "--display-name-prefix, or --tags to "
-                "specify which tests to run."
-            )
-            sys.exit(1)
-
-        if args.display_name_prefix:
-            print(
-                "Fetching tests matching prefix: "
-                f"'{args.display_name_prefix}'..."
-            )
-        elif args.tags:
-            print(f"Fetching tests matching tags: {args.tags}...")
-        all_evals = eval_client.list_evaluations(app_name=args.app_name)
-
-        for eval_obj in all_evals:
-            match = False
-
-            if args.display_name_prefix and eval_obj.display_name.startswith(
-                args.display_name_prefix
-            ):
-                match = True
-
-            # Assuming tags are accessible as a
-            # list/repeated field on the Evaluation
-            # object
-            if args.tags and hasattr(eval_obj, "tags"):
-                # intersection of CLI tags and agent tags
-                if any(t in eval_obj.tags for t in args.tags):
-                    match = True
-
-            if match:
-                evaluations_to_run.append(eval_obj.name)
-
-        if not evaluations_to_run:
-            print(
-                "No matching tests found for the "
-                "given prefix or tags. Aborting run."
-            )
-            sys.exit(0)
-
-        print(f"Found {len(evaluations_to_run)} matching test(s) to run.")
-
-    # Determine which evaluations to run
-    evaluations_to_run = []
-    if args.evaluation_id:
-        evaluations_to_run.append(args.evaluation_id)
-    else:
-        # Require prefix or tags if no specific ID is given
-        if not args.display_name_prefix and not args.tags:
-            print(
-                "Error: You must provide either --evaluation-id, "
-                "--display-name-prefix, or --tags to "
-                "specify which tests to run."
-            )
-            sys.exit(1)
-
-        if args.display_name_prefix:
-            print(
-                "Fetching tests matching prefix: "
-                f"'{args.display_name_prefix}'..."
-            )
-        elif args.tags:
-            print(f"Fetching tests matching tags: {args.tags}...")
-        all_evals = eval_client.list_evaluations(app_name=args.app_name)
-
-        for eval_obj in all_evals:
-            match = False
-
-            if args.display_name_prefix and eval_obj.display_name.startswith(
-                args.display_name_prefix
-            ):
-                match = True
-
-            # Assuming tags are accessible as a
-            # list/repeated field on the Evaluation
-            # object
-            if args.tags and hasattr(eval_obj, "tags"):
-                # intersection of CLI tags and agent tags
-                if any(t in eval_obj.tags for t in args.tags):
-                    match = True
-
-            if match:
-                evaluations_to_run.append(eval_obj.name)
-
-        if not evaluations_to_run:
-            print(
-                "No matching tests found for the "
-                "given prefix or tags. Aborting run."
-            )
-            sys.exit(0)
-
-        print(f"Found {len(evaluations_to_run)} matching test(s) to run.")
-
-    try:
-        # Step 1: Capture existing evaluation runs to diff against later
-        df_initial = eval_utils.evals_to_dataframe().get(
-            "summary", pd.DataFrame()
-        )
-        old_result_ids = set()
-        if not df_initial.empty and "eval_result_id" in df_initial.columns:
-            old_result_ids = set(df_initial["eval_result_id"].unique())
-
-        # Step 2: Trigger evaluation
-        eval_client.run_evaluation(
-            evaluations=evaluations_to_run, app_name=args.app_name
-        )
-        print("Evaluation triggered successfully based on CLI call.")
-
-        # Step 3: Wait and backoff on pending evaluations.
-        if args.wait:
-            df_new_run = wait_for_evaluation_completion(
-                eval_utils,
-                old_result_ids,
-                args.app_name,
-                expected_count=len(evaluations_to_run),
-            )
-            pass_status = filter_metrics_and_assess(
-                df_new_run, args.filter_auto_metrics
-            )
-
-            if pass_status:
-                print("\nFINAL RESULT: PASS")
-                sys.exit(0)
-            else:
-                df_failures = df_new_run.get("failures", pd.DataFrame())
-                if not df_failures.empty:
-                    print("\n--- Failure Details ---")
-                    grouped = df_failures.groupby("display_name", sort=False)
-                    for disp, group_df in grouped:
-                        is_err = any(
-                            row.get("failure_type") == "System Engine Error"
-                            for _, row in group_df.iterrows()
-                        )
-                        title_str = "Errored" if is_err else "Failed"
-                        print(f"\n{disp} {title_str}")
-
-                        sys_errors = group_df[
-                            group_df["failure_type"] == "System Engine Error"
-                        ]
-                        normal_fails = group_df[
-                            group_df["failure_type"] != "System Engine Error"
-                        ]
-
-                        for _, row in sys_errors.iterrows():
-                            print(f"- {row.get('actual')}\n")
-
-                        for _, row in normal_fails.iterrows():
-                            idx = row.get("turn_index")
-                            tba = f" (Turn {idx})" if pd.notnull(idx) else ""
-
-                            print(f"- Type    : {row.get('failure_type')}{tba}")
-                            print(f"- Expected: {row.get('expected')}")
-                            print(f"- Actual  : {row.get('actual')}")
-
-                            score = row.get("score")
-                            if pd.notnull(score):
-                                print(f"- Score   : {score}")
-                            print()
-
-                print("\nFINAL RESULT: FAIL")
-                sys.exit(1)
-    except Exception as e:
-        print(f"Failed to run evaluation: {e}")
-        sys.exit(1)
-
-
-def combined_evals_report_cmd(args: argparse.Namespace) -> None:
-    """Handles the 'evals report' command."""
-    from cxas_scrapi.utils.reporting import (  # noqa: PLC0415
-        generate_combined_report_from_dir,
-    )
-
-    output_path = (
-        args.gcs_path
-        or args.output
-        or os.path.join(args.output_dir, "combined_report.html")
-    )
-
-    include_list = args.include.split(",") if args.include else []
-    filter_files_list = (
-        args.filter_files.split(",")
-        if getattr(args, "filter_files", None)
-        else []
-    )
-    filter_tags_list = (
-        args.filter_tags.split(",")
-        if getattr(args, "filter_tags", None)
-        else []
-    )
-
-    if getattr(args, "input_dir", None):
-        if args.tool_test_file == "evals/tool_tests/":
-            args.tool_test_file = os.path.join(args.input_dir, "tool_tests/")
-        if args.goldens_dir == "evals/goldens/":
-            args.goldens_dir = os.path.join(args.input_dir, "goldens/")
-        if args.simulation_dir == "evals/simulations/":
-            args.simulation_dir = os.path.join(args.input_dir, "simulations/")
-
-    sim_parallel = getattr(args, "sim_parallel", 5)
-    golden_timeout = getattr(args, "golden_timeout", 600)
-
-    generate_combined_report_from_dir(
-        output_dir=args.output_dir,
-        golden_run=args.golden_run,
-        app_name=args.app_name,
-        output_path=output_path,
-        run=args.run,
-        app_dir=args.app_dir,
-        tool_test_file=args.tool_test_file,
-        goldens_dir=args.goldens_dir,
-        simulation_dir=args.simulation_dir,
-        format=args.format,
-        include=include_list,
-        modality=args.modality,
-        runs=args.runs,
-        filter_files=filter_files_list,
-        filter_tags=filter_tags_list,
-        parallel=sim_parallel,
-        golden_timeout=golden_timeout,
-        bg_noise_file=getattr(args, "bg_noise_file", None),
-        burst_noise_files=getattr(args, "burst_noise_files", "").split(",")
-        if getattr(args, "burst_noise_files", None)
-        else None,
-    )
-    print(f"Combined report generated at {output_path}")
-
-
-def test_tools(args: argparse.Namespace) -> None:
-    """Handles the 'test-tools' command."""
-    from cxas_scrapi.evals.tool_evals import ToolEvals
-
-    print(
-        f"Running tool tests for App: {args.app_name} "
-        f"using file: {args.test_file}"
-    )
-    tool_evals = ToolEvals(app_name=args.app_name)
-
-    try:
-        test_cases = tool_evals.load_tool_test_cases_from_file(args.test_file)
-        if not test_cases:
-            print(f"No valid test cases found in {args.test_file}")
-            sys.exit(1)
-
-        results = tool_evals.run_tool_tests(test_cases, debug=args.debug)
-
-        # Check overall status
-        failed_count = sum(1 for r in results["status"] if r != "PASSED")
-
-        if failed_count > 0:
-            print(f"\nFINAL RESULT: FAIL ({failed_count} tools failed)")
-            sys.exit(1)
-        else:
-            print(f"\nFINAL RESULT: PASS (All {len(results)} tools passed)")
-            sys.exit(0)
-
-    except Exception as e:
-        print(f"Failed to run tool tests: {e}")
-        sys.exit(1)
-
-
-def test_callbacks(args: argparse.Namespace) -> None:
-    """Handles the 'test-callbacks' command."""
-    from cxas_scrapi.evals.callback_evals import CallbackEvals
-
-    print(f"Running callback tests in App directory: {args.app_dir}")
-    callback_evals = CallbackEvals()
-
-    try:
-        results = callback_evals.test_all_callbacks_in_app_dir(
-            app_dir=args.app_dir,
-            agent_name=args.agent_name,
-            callback_type=args.callback_type,
-            callback_name=args.callback_name,
-            log_file=args.log_file,
-            pytest_args=args.pytest_args,
-        )
-        if results.empty:
-            print(f"No valid callback tests found in {args.app_dir}")
-            sys.exit(1)
-
-        # Check overall status
-        failed_count = sum(1 for r in results["status"] if r != "PASSED")
-
-        if failed_count > 0:
-            print(f"\nFINAL RESULT: FAIL ({failed_count} callbacks failed)")
-            sys.exit(1)
-        else:
-            print(f"\nFINAL RESULT: PASS (All {len(results)} callbacks passed)")
-            sys.exit(0)
-
-    except Exception as e:
-        print(f"Failed to run callback tests: {e}")
-        sys.exit(1)
-
-
-def test_single_callback(args: argparse.Namespace) -> None:
-    """Handles the 'test-single-callback' command."""
-    from cxas_scrapi.evals.callback_evals import CallbackEvals
-
-    print(
-        f"Running single callback test for "
-        f"Agent: {args.agent_name}, "
-        f"Type: {args.callback_type}"
-    )
-    callback_evals = CallbackEvals()
-
-    try:
-        results = callback_evals.test_single_callback_for_agent(
-            app_name=args.app_name,
-            agent_name=args.agent_name,
-            callback_type=args.callback_type,
-            test_file_path=args.test_file_path,
-            log_file=args.log_file,
-            pytest_args=args.pytest_args,
-        )
-        if results.empty:
-            print(f"No valid callback tests found at {args.test_file_path}")
-            sys.exit(1)
-
-        # Check overall status
-        failed_count = sum(1 for r in results["status"] if r != "PASSED")
-
-        if failed_count > 0:
-            print(f"\nFINAL RESULT: FAIL ({failed_count} callbacks failed)")
-            sys.exit(1)
-        else:
-            print(f"\nFINAL RESULT: PASS (All {len(results)} callbacks passed)")
-            sys.exit(0)
-
-    except Exception as e:
-        print(f"Failed to run callback tests: {e}")
-        sys.exit(1)
-
-
-def ci_test(args: argparse.Namespace) -> None:
-    """Handles the 'ci-test' command."""
-    from cxas_scrapi.cli.app import app_push
-    from cxas_scrapi.core.apps import Apps
-    from cxas_scrapi.core.evaluations import Evaluations
-
-    print("Starting CI Test Lifecycle...")
-
-    if hasattr(args, "display_name") and args.display_name:
-        temp_display_name = args.display_name
-    else:
-        temp_display_name = f"[CI] PR Test {uuid.uuid4().hex[:8]}"
-
-    args.display_name = temp_display_name
-    args.app_name = None  # Force create by default
-
-    apps_client = Apps(project_id=args.project_id, location=args.location)
-
-    existing_app = apps_client.get_app_by_display_name(temp_display_name)
-    if existing_app:
-        print(f"Found existing temp agent: {existing_app.name}. Updating...")
-        args.app_name = existing_app.name
-
-    temp_app_name = app_push(args)
-
-    if not temp_app_name:
-        print("Failed to get deployed temp app name. CI Test aborting.")
-        sys.exit(1)
-
-    try:
-        # Run test-tools
-
-        test_file = os.path.join(args.app_dir, "tests", "tool_tests.yaml")
-        if os.path.exists(test_file):
-            print(f"\\n--- Running Tool Tests on {temp_app_name} ---")
-            cmd = [
-                "cxas",
-                "test-tools",
-                "--app-name",
-                temp_app_name,
-                "--test-file",
-                test_file,
-            ]
-            print(f"Executing: {' '.join(cmd)}")
-            res = subprocess.run(cmd, check=False)
-            if res.returncode != 0:
-                print("Tool tests failed.")
-                sys.exit(1)
-
-        # We must evaluate using the API or SDK
-        print(f"\\n--- Running Evaluations on {temp_app_name} ---")
-
-        evals_client = Evaluations(app_name=temp_app_name)
-        evals_map = evals_client.get_evaluations_map()
-
-        if not evals_map or (
-            not evals_map.get("goldens") and not evals_map.get("scenarios")
-        ):
-            print("No evaluations found in the temp app. Skipping run_eval.")
-        else:
-            all_eval_ids = list(evals_map.get("goldens", {}).values()) + list(
-                evals_map.get("scenarios", {}).values()
-            )
-            for eval_id in all_eval_ids:
-                cmd = [
-                    "cxas",
-                    "run",
-                    "--app-name",
-                    temp_app_name,
-                    "--evaluation-id",
-                    eval_id,
-                    "--wait",
-                    "--filter-auto-metrics",
-                ]
-                print(f"Executing: {' '.join(cmd)}")
-                res = subprocess.run(cmd, check=False)
-                if res.returncode != 0:
-                    print(f"Evaluation '{eval_id}' failed.")
-                    sys.exit(1)
-
-        print(
-            "\\nCI Test Lifecycle Completed Successfully! "
-            "Temp agent persists for review."
-        )
-
-    except Exception as e:
-        print(f"Failed to execute CI Tests: {e}")
-        sys.exit(1)
-
-
-def local_test(args: argparse.Namespace) -> None:
-    """Handles the 'local-test' command."""
-
-    agent_dir = os.path.abspath(args.app_dir)
-    agent_name = (
-        os.path.basename(agent_dir.rstrip(os.sep)).lower().replace(" ", "-")
-    )
-    tag = f"{agent_name}-local-test"
-
-    print(f"Building Docker image for {agent_name}...")
-    # Compilation requires executing from the root agent directory
-    build_cmd = ["docker", "build", "-t", tag, agent_dir]
-    if subprocess.call(build_cmd) != 0:
-        print("Docker build failed.")
-        sys.exit(1)
-
-    print("Running tests in Docker container...")
-
-    # Detect ADC
-    home = os.path.expanduser("~")
-    # Default gcloud location
-    adc_path = os.environ.get("GOOGLE_APPLICATION_CREDENTIALS")
-    if not adc_path:
-        adc_path = os.path.join(
-            home, ".config/gcloud/application_default_credentials.json"
-        )
-
-    docker_cmd = [
-        "docker",
-        "run",
-        "--rm",
-        "-v",
-        f"{agent_dir}:/workspace",
-        "-w",
-        "/workspace",
-        "-e",
-        f"PROJECT_ID={args.project_id}",
-        "-e",
-        f"LOCATION={args.location}",
-    ]
-
-    oauth_token = os.environ.get("CXAS_OAUTH_TOKEN")
-
-    if oauth_token:
-        print("Using provided CXAS_OAUTH_TOKEN.")
-        docker_cmd.extend(["-e", "CXAS_OAUTH_TOKEN"])
-    elif os.path.exists(adc_path):
-        print(f"Mounting credentials from {adc_path}")
-        docker_cmd.extend(
-            [
-                "-e",
-                "GOOGLE_APPLICATION_CREDENTIALS=/tmp/keys/adc.json",
-                "-v",
-                f"{adc_path}:/tmp/keys/adc.json:ro",
-            ]
-        )
-    else:
-        print(
-            "Warning: Application Default Credentials not found. "
-            "Authentication may fail."
-        )
-
-    display_name = f"[Local] {agent_name}"
-
-    # The command passed to the container
-    inner_cmd = [
-        tag,
-        "ci-test",
-        "--app-dir",
-        "/workspace",
-        "--project-id",
-        args.project_id,
-        "--location",
-        args.location,
-        "--display-name",
-        display_name,
-    ]
-
-    env_file = getattr(args, "env_file", None)
-    if env_file:
-        inner_cmd.extend(["--env-file", env_file])
-
-    docker_cmd.extend(inner_cmd)
-
-    print(f"Executing: {' '.join(docker_cmd)}")
-    sys.exit(subprocess.call(docker_cmd))
-
-
-def run_session(args: argparse.Namespace) -> None:
-    """Handles the 'run-session' command."""
-    from cxas_scrapi import Sessions
-
-    try:
-        session_client = Sessions(args.app_name)
-        session_id = session_client.create_session_id()
-
-        while True:
-            try:
-                user_input = input()
-            except (EOFError, KeyboardInterrupt):
-                break
-
-            if not user_input.strip():
-                continue
-
-            res = session_client.run(
-                session_id=session_id,
-                text=user_input,
-                modality=args.modality,
-                use_tool_fakes=args.use_tool_fakes,
-            )
-            session_client.parse_result(res)
-    except Exception as e:
-        print(f"Failed to run session: {e}")
-        sys.exit(1)
-
-
-def conversations_list(args: argparse.Namespace) -> None:
-    """Lists conversations for an app."""
-    from google.protobuf.json_format import MessageToDict
-
-    from cxas_scrapi.core.apps import Apps
-    from cxas_scrapi.core.common import Common
-    from cxas_scrapi.core.conversation_history import ConversationHistory
-
-    print(f"Listing conversations for App: {args.app_name}")
-
-    # Extract and validate app_name
-    app_name = Common._get_app_name(args.app_name)
-    if not app_name:
-        print(
-            "Error: Invalid App Name format. Please use the full resource "
-            "name in the format 'projects/.../locations/.../apps/...'"
-        )
-        sys.exit(1)
-
-    project_id = Common._get_project_id(args.app_name)
-    location = Common._get_location(args.app_name)
-
-    apps_client = Apps(project_id=project_id, location=location)
-    ch_client = ConversationHistory(
-        app_name=args.app_name, creds=apps_client.creds
-    )
-    conversations = ch_client.list_conversations()
-
-    conversations_dict = []
-    for c in conversations:
-        try:
-            c_dict = MessageToDict(c._pb)
-        except AttributeError:
-            c_dict = MessageToDict(c)
-        conversations_dict.append(c_dict)
-
-    print(json.dumps(conversations_dict, indent=2))
-
-
-def conversations_get(args: argparse.Namespace) -> None:
-    """Gets details of a specific conversation."""
-    from google.protobuf.json_format import MessageToDict
-
-    from cxas_scrapi.core.apps import Apps
-    from cxas_scrapi.core.common import Common
-    from cxas_scrapi.core.conversation_history import ConversationHistory
-
-    print(f"Getting conversation: {args.conversation_resource_name}")
-
-    # Extract and validate app_name
-    app_name = Common._get_app_name(args.conversation_resource_name)
-    if not app_name:
-        print(
-            "Error: Invalid Conversation Resource Name format. Please use the "
-            "full resource name in the format "
-            "'projects/.../locations/.../apps/.../conversations/...'"
-        )
-        sys.exit(1)
-
-    project_id = Common._get_project_id(args.conversation_resource_name)
-    location = Common._get_location(args.conversation_resource_name)
-
-    apps_client = Apps(project_id=project_id, location=location)
-    ch_client = ConversationHistory(app_name=app_name, creds=apps_client.creds)
-    conv = ch_client.get_conversation(
-        conversation_id=args.conversation_resource_name
-    )
-
-    try:
-        conv_dict = MessageToDict(conv._pb)
-    except AttributeError:
-        conv_dict = MessageToDict(conv)
-
-    print(json.dumps(conv_dict, indent=2))
-
-
-def deployments_list(args: argparse.Namespace) -> None:
-    """Lists deployments for an app."""
-    from google.protobuf.json_format import MessageToDict
-
-    from cxas_scrapi.core.deployments import Deployments
-
-    print(f"Listing deployments for App: {args.app_name}")
-
-    deployments_client = Deployments(app_name=args.app_name)
-    deployments = deployments_client.list_deployments()
-
-    try:
-        deployments_dict = []
-        for d in deployments:
-            try:
-                d_dict = MessageToDict(d._pb)
-            except AttributeError:
-                d_dict = MessageToDict(d)
-            deployments_dict.append(d_dict)
-        print(json.dumps(deployments_dict, indent=2))
-    except Exception:
-        # Fallback to string representation if serialization fails
-        print([str(d) for d in deployments])
-
-
-def deployments_create(args: argparse.Namespace) -> None:
-    """Creates a deployment."""
-    from cxas_scrapi.core.deployments import Deployments
-
-    print(f"Creating deployment {args.deployment_id} for App: {args.app_name}")
-
-    deployments_client = Deployments(app_name=args.app_name)
-    deployment = deployments_client.create_deployment(
-        deployment_id=args.deployment_id,
-        display_name=args.deployment_id,
-        app_version=args.version_id,
-    )
-    print(f"Deployment created successfully: {deployment.name}")
-
-
-def deployments_promote(args: argparse.Namespace) -> None:
-    """Promotes app to live traffic."""
-    from google.api_core.exceptions import NotFound
-
-    from cxas_scrapi.cli.app import app_push
-    from cxas_scrapi.core.deployments import Deployments
-
-    print(f"Promoting app {args.app_resource_name} to live traffic...")
-
-    # Step 1: Push and create version
-    push_args = argparse.Namespace(
-        app_dir=args.app_dir,
-        to=args.app_resource_name,
-        app_name=None,
-        display_name=None,
-        env_file=None,
-        project_id=None,
-        location=None,
-        create_version=True,
-        version_description=f"Promote {time.strftime('%Y%m%d%H%M%S')}",
-    )
-
-    try:
-        print("Calling app_push directly...")
-        app_name = app_push(push_args)
-        if not app_name:
-            print("Error: Push failed during promotion.")
-            sys.exit(1)
-
-        version_id = getattr(push_args, "created_version_name", None)
-        if not version_id:
-            print("Error: Could not get created version ID.")
-            sys.exit(1)
-
-        # Step 2: Update deployment
-        deployment_id = args.live_deployment_resource_name.split(
-            "/deployments/"
-        )[-1]
-
-        deployments_client = Deployments(app_name=args.app_resource_name)
-
-        try:
-            deployments_client.get_deployment(deployment_id=deployment_id)
-        except NotFound:
-            print(f"Error: Deployment '{deployment_id}' does not exist.")
-            print(
-                "`deployments promote` requires "
-                "promoting an existing deployment."
-            )
-            print(
-                "Please create the deployment first using `deployments create`."
-            )
-            sys.exit(1)
-
-        print(
-            f"Updating deployment {deployment_id} with version {version_id}..."
-        )
-        deployments_client.update_deployment(
-            deployment_id=deployment_id, app_version=version_id
-        )
-
-        print("Successfully promoted agent to live traffic.")
-
-    except Exception as e:
-        print(f"Error during promotion: {e}")
-        sys.exit(1)
 
 
 def get_parser() -> argparse.ArgumentParser:
